pyproject.toml
[project]
name = "mcprotocol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.1",
    "langgraph>=1.0.9",
    "mcp>=1.26.0",
    "fastapi",
    "uvicorn[standard]",
    "python-dotenv",
    "langchain-community",
]
